pyproject.toml
[tool.poetry]
name = "interview"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.1.1"
langchain = "^0.3.26"
langchain-community = "^0.3.27"
pyowm = "^3.3.0"
setuptools = "<70"
langgraph = "^0.5.1"
langchain-openai = "^0.3.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
